Added docblock comments. Restructured page objects
