+++ packages/flex-plugin-e2e-tests/src/utils/pages/view/agent-desktop.ts
@@ -5,6 +5,9 @@ import { Base } from '../base';
 export class AgentDesktop extends Base {
   // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
   assert = {
+    /**
+     * Checks whether task canvas are visible on the Agent Desktop
+     */
     isVisible: async (): Promise<ElementHandle<Element>> => this.elementVisible(this._noTaskCanvas, 'Task canvas'),
   };
 
@@ -17,6 +20,9 @@ export class AgentDesktop extends Base {
     this._baseUrl = baseUrl;
   }
 
+  /**
+   * Navigates to agent-desktop
+   */
   async open(): Promise<void> {
     await this.goto({ baseUrl: this._baseUrl, path: 'agent-desktop' });
   }
